Move ConfigLoader behind the embulk-deps DependencyClassLoader
